Refactoring error notification into a base class for ViewModels
